web/back/requirements.txt
fastapi==0.115.13
uvicorn==0.34.3
pydantic-settings==2.10.1
SQLAlchemy==2.0.41
redis==6.2.0
orjson==3.10.18
pyJWT==2.10.1
httpx==0.28.1
backoff==2.2.1
werkzeug==3.1.3

alembic==1.16.2
asyncpg==0.30.0
psycopg2-binary==2.9.10